refactor:LottoUtil 메서드 추가
